requirements.txt
flask
gunicorn
google-auth
google-auth-httplib2
google-api-python-client
google-ads>=24.0.0
pyyaml
requests
